# contracts/http/openapi.yaml
openapi: 3.1.0
info:
  title: VPN Subscription Platform API
  version: 0.7.1-stage7
  license:
    name: UNLICENSED
    identifier: LicenseRef-UNLICENSED
  description: >
    Stage 7 contract. Durable Telegram notification delivery and the bounded
    mTLS administrator API/CLI are executable locally. Production deployment
    and credential issuance remain out of scope.
servers:
  - url: https://api.example.invalid
    description: Public API placeholder
  - url: https://internal.example.invalid
    description: Internal API placeholder requiring mTLS service identity
  - url: https://admin.example.invalid
    description: Admin API placeholder requiring admin mTLS identity
  - url: https://sub.example.invalid
    description: Happ subscription endpoint placeholder
security:
  - ServiceMTLS: []
...
paths:
...
  /internal/v1/telegram-users/{telegram_id}:
    put:
      tags: [Identity]
      summary: Idempotently register or refresh a Telegram identity
      operationId: upsertTelegramIdentity
      x-contract-status: implemented-stage2
      security:
        - ServiceMTLS: []
      x-allowed-service-identities: [telegram-bot]
      parameters:
        - $ref: '#/components/parameters/RequestId'
        - name: telegram_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TelegramIdentityUpsertRequest'
      responses:
        '200':
          description: Existing or created identity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserIdentity'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
...
  /internal/v1/users/{user_id}/notification-target:
    get:
      tags: [Identity]
      summary: Resolve a consent-eligible Telegram delivery target
      operationId: getNotificationTarget
      x-contract-status: implemented-stage7
      x-allowed-service-identities: [notification-service]
      parameters:
        - $ref: '#/components/parameters/UserId'
        - name: document_type
          in: query
          required: true
          schema: {type: string, minLength: 1, maxLength: 64}
        - name: document_version
          in: query
          required: true
          schema: {type: string, minLength: 1, maxLength: 64}
      responses:
        '200':
          description: Eligibility result; chat ID is returned only to notification-service
          content:
            application/json:
              schema: {$ref: '#/components/schemas/NotificationTarget'}
        '404': {$ref: '#/components/responses/Error'}
        default: {$ref: '#/components/responses/Error'}
...
components:
  securitySchemes:
    ServiceMTLS:
      type: mutualTLS
      description: >
        Internal service identity from a trusted platform CA. Authorization is
        based on a strict SPIFFE URI from the verified certificate chain and
        per-endpoint allowed identities.
    ProvisioningMTLS:
      type: mutualTLS
      description: >
        Internal mTLS identity restricted to provisioning-service for fetching
        sensitive credential material from access-service.
    NodeAgentMTLS:
      type: mutualTLS
      description: >
        Private management-plane mTLS. The client must be provisioning-service,
        and provisioning-service additionally verifies the exact registered
        node SPIFFE URI in the server's verified certificate chain.
    AdminMTLS:
      type: mutualTLS
      description: >
        Admin CLI/client mTLS identity mapped to RBAC roles and audited actions.
    TelegramWebhookSecret:
      type: apiKey
      in: header
      name: X-Telegram-Bot-Api-Secret-Token
      description: Telegram webhook secret token configured through Bot API setWebhook.
    SubscriptionUrlBearer:
      type: http
      scheme: bearer
      bearerFormat: URL path token
      description: >
        Bearer-secret semantics for `GET /s/{token}`. Happ v1 carries this
        token in the path, not in the Authorization header.
  parameters:
    RequestId:
      name: X-Request-Id
      in: header
      required: false
      schema:
        type: string
        minLength: 8
        maxLength: 128
...
    UserId:
      name: user_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    Error:
      description: Safe error response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      required: [code, message, request_id]
      additionalProperties: false
      properties:
        code:
          type: string
          maxLength: 128
        message:
          type: string
          maxLength: 512
        request_id:
          type: string
          maxLength: 128
        fields:
          type: array
          items:
            type: object
            required: [field, code]
            additionalProperties: false
            properties:
              field:
                type: string
              code:
                type: string
              message:
                type: string
    UserIdentity:
      type: object
      required: [user_id, status]
      additionalProperties: false
      properties:
        user_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [active, blocked, deleted]
        locale:
          type: [string, 'null']
          maxLength: 32
        timezone:
          type: [string, 'null']
          maxLength: 64
    TelegramIdentityUpsertRequest:
      type: object
      additionalProperties: false
      properties:
        username:
          type: [string, 'null']
          maxLength: 64
        display_name:
          type: [string, 'null']
          maxLength: 256
        language_code:
          type: [string, 'null']
          maxLength: 16
        locale:
          type: [string, 'null']
          maxLength: 32
...
    NotificationTarget:
      type: object
      required: [eligible]
      additionalProperties: false
      properties:
        eligible: {type: boolean}
        reason_code: {type: string, maxLength: 64}
        telegram_chat_id: {type: integer, format: int64, minimum: 1, writeOnly: true}
        locale: {type: string, maxLength: 32}
...